fix: Fix RoomChattingUI
