+++ Assets/Resources/Chatter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d6af3644b2020ea41a8f49509ddbb132
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
