+++ Assets/Resources/UI/GameStartUI.prefab
@@ -485,9 +485,9 @@ RectTransform:
   m_Father: {fileID: 3648915761824057080}
   m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 960, y: 540}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 1920, y: 1080}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!223 &3648915759921990171
@@ -7290,9 +7290,9 @@ RectTransform:
   m_Father: {fileID: 3648915761824057080}
   m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 960, y: 540}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 1920, y: 1080}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!223 &3648915761666700066
@@ -9102,9 +9102,7 @@ PrefabInstance:
     - target: {fileID: 5927651940847589453, guid: faf72926bc734ef4ebf2e8a85d566758,
         type: 3}
       propertyPath: m_text
-      value: 'SIGN IN
-
-'
+      value: SIGN IN
       objectReference: {fileID: 0}
     - target: {fileID: 7789668459027263726, guid: faf72926bc734ef4ebf2e8a85d566758,
         type: 3}
