+++ Assets/Resources/UI/InGameUI.prefab
@@ -960,7 +960,6 @@ GameObject:
   - component: {fileID: 4243693414673562209}
   - component: {fileID: 4243693414673562214}
   - component: {fileID: 4243693414673562215}
-  - component: {fileID: 2088749142}
   m_Layer: 5
   m_Name: Chatting
   m_TagString: Untagged
@@ -1059,20 +1058,6 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
---- !u!114 &2088749142
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4243693414673562213}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e322fdbef3a62ef489c782849cda821e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chatInput: {fileID: 4243693415600209698}
-  content: {fileID: 4243693414037496500}
 --- !u!1 &4243693414762214254
 GameObject:
   m_ObjectHideFlags: 0
