+++ Assets/Scripts/JBB/UI/Chatter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 091d8c49a5cb3f441a9d05eaa37c5dca
+guid: 157bd1e7625129e4b9259b1a33988786
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
