+++ Assets/Scripts/JBB/UI/RoomChattingUI.cs
@@ -8,16 +8,16 @@ namespace JBB
 {
     public class RoomChattingUI : BaseUI
     {
+        public TMP_InputField chatInput;
+        public GameObject content;
         protected override void Awake()
         {
-            base.Awake();
-            
+            base.Awake();            
         }
         private void Start()
         {
-            GameObject go = PhotonNetwork.Instantiate("UI/RoomChattingPrefab", transform.position, transform.rotation);
-            go.transform.SetParent(transform, false);
-            go.transform.SetPositionAndRotation(transform.position, transform.rotation);
+            GameObject go = PhotonNetwork.Instantiate("Chatter", Vector3.zero, Quaternion.identity);
+            Chatter chatter = go.GetComponent<Chatter>();
         }
 
     }
